Moved quest registration over
