+++ src/WarcraftLegacies.Source/Factions/FelHorde.cs
@@ -2,6 +2,7 @@
 using MacroTools;
 using MacroTools.FactionSystem;
 using WarcraftLegacies.Source.FactionMechanics.Fel_Horde;
+using WarcraftLegacies.Source.Quests.Fel_Horde;
 using WarcraftLegacies.Source.Setup;
 using WCSharp.Shared.Data;
 using static War3Api.Common;
@@ -135,7 +136,19 @@ The Alliance is gathering outside the Dark Portal to stop you, so prepare to for
 
     private void RegisterQuests()
     {
-      throw new System.NotImplementedException();
+      var questHellfireCitadel = FelHorde.AddQuest(new QuestHellfireCitadel(Regions.HellfireUnlock));
+      FelHorde.AddQuest(new QuestRuinsofShadowmoon(Regions.ShadowmoonBaseUnlock));
+      FelHorde.AddQuest(new QuestDarkPortal(
+        preplacedUnitSystem.GetUnit(Constants.UNIT_N036_DARK_PORTAL_WAYGATE, Regions.Dark_Portal_Entrance_1.Center),
+        preplacedUnitSystem.GetUnit(Constants.UNIT_N036_DARK_PORTAL_WAYGATE, Regions.Dark_Portal_Entrance_2.Center),
+        preplacedUnitSystem.GetUnit(Constants.UNIT_N036_DARK_PORTAL_WAYGATE, Regions.Dark_Portal_Entrance_3.Center),
+        preplacedUnitSystem.GetUnit(Constants.UNIT_N036_DARK_PORTAL_WAYGATE, Regions.Dark_Portal_Exit_1.Center),
+        preplacedUnitSystem.GetUnit(Constants.UNIT_N036_DARK_PORTAL_WAYGATE, Regions.Dark_Portal_Exit_2.Center),
+        preplacedUnitSystem.GetUnit(Constants.UNIT_N036_DARK_PORTAL_WAYGATE, Regions.Dark_Portal_Exit_3.Center)));
+      FelHorde.AddQuest(new QuestBlackrock(Regions.BlackrockUnlock, Regions.DarkPortalUnlock, new[] { questHellfireCitadel }));
+      FelHorde.AddQuest(new QuestFelHordeKillIronforge(allLegendSetup.Ironforge.GreatForge));
+      FelHorde.AddQuest(new QuestFelHordeKillStormwind(allLegendSetup.Stormwind.StormwindKeep));
+      FelHorde.AddQuest(new QuestGuldansLegacy());
     }
   }
 }
