+++ src/WarcraftLegacies.Source/Setup/SharedQuestSetup.cs
